created a base persistence class with shared methods, create separate persistence classes for main persistence and parent user folder
